+++ app/src/main/java/com/mtcnextlabs/imnuricrestine/models/FavoritesListItem.kt
@@ -1,8 +1,8 @@
 package com.mtcnextlabs.imnuricrestine.models
 
 data class FavoritesListItem(
-    val id: Short,
-    val hymnId: Short,
+    val id: Int,
+    val hymnId: Int,
     val index: String,
     val title: String
 )
